refactor(services): move SanitizeProjectName to projectenv
