+++ internal/postgres/envvars.go
@@ -3,7 +3,7 @@ package postgres
 import "fmt"
 
 // EnvVars returns the DB_* map injected into a linked project's .env.
-// projectName is sanitized by the caller (services.SanitizeProjectName).
+// projectName is sanitized by the caller (projectenv.SanitizeProjectName).
 func EnvVars(projectName, major string) (map[string]string, error) {
 	port, err := PortFor(major)
 	if err != nil {
